requirements.txt
Django==4.1.9
djangorestframework==3.14.0
opencv-python-headless==4.7.0.72
pandas==1.5.3
Pillow==9.5.0
torch==2.0.1
ultralytics==8.0.114
python-dotenv==1.0.0
